docs: move getRoutes to utils
